fix(issue-433): Address PR review - split hook lifecycle tests
