+++ cli/src/installer/hooks.rs
@@ -4,6 +4,11 @@ use crate::hook::Hook;
 use anyhow::{Context, Result};
 use std::path::{Path, PathBuf};
 
+mod opencode;
+
+use opencode::{install_hook_opencode, remove_hook_from_opencode_json};
+pub(crate) use opencode::{opencode_hook_instruction_contents, opencode_hook_instruction_path};
+
 pub(crate) fn validate_item_name(name: &str) -> Result<()> {
     if name.is_empty() {
         anyhow::bail!("item name must not be empty");
@@ -140,8 +145,8 @@ fn claude_owned_hook_commands(global: bool, hook_name: &str, script_path: &Path)
 /// - OpenCode: add permission rules to opencode.json
 /// - Codex: native hooks.json entry + script when codex supports the event;
 ///   safety prose appended to agent TOML developer_instructions otherwise
-/// - Cursor: append safety advisory to all .mdc rule files
-/// - Pi: no-op (safety prose lives in agent bodies via the Pi generator)
+/// - Cursor: write a dedicated `.cursor/rules/safety-<name>.mdc` rule file
+/// - Pi: no per-hook install artifact; the native `pi-hooks` extension owns behavior
 ///
 /// Honors the optional `harnesses:` allowlist in the hook frontmatter.
 pub fn install_hook(
@@ -259,51 +264,6 @@ fn install_hook_claude(hook: &Hook, global: bool) -> Result<()> {
     Ok(())
 }
 
-/// OpenCode: add permission rules based on hook intent
-fn install_hook_opencode(hook: &Hook, global: bool) -> Result<()> {
-    validate_item_name(&hook.name)?;
-    let config_path = if global {
-        crate::config::opencode_global_config_path()
-    } else {
-        crate::config::opencode_project_config_path()
-    };
-    let instruction_dir = opencode_hook_instruction_dir(global);
-    if instruction_dir.exists() {
-        checked_child_path(&instruction_dir, &format!("vstack-hook-{}.md", hook.name))?;
-    }
-    let instruction_path = opencode_hook_instruction_path(global, &hook.name);
-    let instruction_ref = opencode_hook_instruction_ref(global, &hook.name);
-    install_hook_opencode_at_path(hook, &config_path, &instruction_path, &instruction_ref)
-}
-
-fn opencode_hook_instruction_dir(global: bool) -> PathBuf {
-    if global {
-        crate::config::opencode_global_dir().join("instructions")
-    } else {
-        crate::config::project_root()
-            .join(".opencode")
-            .join("instructions")
-    }
-}
-
-pub(crate) fn opencode_hook_instruction_path(global: bool, name: &str) -> PathBuf {
-    let file_name = format!("vstack-hook-{name}.md");
-    opencode_hook_instruction_dir(global).join(file_name)
-}
-
-fn opencode_hook_instruction_ref(global: bool, name: &str) -> String {
-    let file_name = format!("vstack-hook-{name}.md");
-    if global {
-        format!("instructions/{file_name}")
-    } else {
-        format!(".opencode/instructions/{file_name}")
-    }
-}
-
-pub(crate) fn opencode_hook_instruction_contents(hook: &Hook) -> String {
-    format!("# Safety: {}\n\n{}", hook.name, hook.safety_prose())
-}
-
 pub(crate) fn cursor_hook_rule_contents(hook: &Hook) -> String {
     let mut output = String::new();
     output.push_str("---\n");
@@ -320,72 +280,6 @@ pub(crate) fn cursor_hook_rule_contents(hook: &Hook) -> String {
 
 pub(crate) fn codex_hook_safety_block(hook: &Hook) -> String {
     format!("## Safety: {}\n\n{}", hook.name, hook.safety_prose())
-}
-
-fn install_hook_opencode_at_path(
-    hook: &Hook,
-    config_path: &Path,
-    instruction_path: &Path,
-    instruction_ref: &str,
-) -> Result<()> {
-    if let Some(parent) = config_path.parent() {
-        std::fs::create_dir_all(parent)?;
-    }
-    if let Some(parent) = instruction_path.parent() {
-        std::fs::create_dir_all(parent)?;
-        let file_name = instruction_path
-            .file_name()
-            .and_then(|name| name.to_str())
-            .context("OpenCode hook instruction path missing file name")?;
-        checked_child_path(parent, file_name)?;
-    }
-
-    std::fs::write(instruction_path, opencode_hook_instruction_contents(hook))?;
-
-    let mut config: serde_json::Value = if config_path.exists() {
-        let content = std::fs::read_to_string(config_path)?;
-        serde_json::from_str(&content)?
-    } else {
-        serde_json::json!({ "$schema": "https://opencode.ai/config.json" })
-    };
-
-    let map = config.as_object_mut().unwrap();
-
-    // OpenCode doesn't have hooks — convert to permission rules and instructions
-    if !map.contains_key("permission") {
-        map.insert("permission".into(), serde_json::json!({}));
-    }
-
-    // Add safety-relevant permission restrictions based on hook type
-    if hook.event == "PreToolUse" {
-        let perms = map.get_mut("permission").unwrap().as_object_mut().unwrap();
-
-        if hook.matcher.as_deref() == Some("Bash") {
-            // For bash hooks: set bash permission to "ask" (require confirmation)
-            if !perms.contains_key("bash") {
-                perms.insert("bash".into(), serde_json::json!({ "*": "ask" }));
-            }
-        }
-    }
-
-    // OpenCode instructions are file paths, so write a dedicated file and reference it.
-    if !map.contains_key("instructions") {
-        map.insert("instructions".into(), serde_json::json!([]));
-    }
-    let instructions = map.get_mut("instructions").unwrap().as_array_mut().unwrap();
-
-    let already_has = instructions
-        .iter()
-        .any(|i| i.as_str() == Some(instruction_ref));
-
-    if !already_has {
-        instructions.push(serde_json::Value::String(instruction_ref.to_string()));
-    }
-
-    let output = serde_json::to_string_pretty(&config)?;
-    std::fs::write(config_path, output)?;
-
-    Ok(())
 }
 
 /// Map a canonical (Claude-style) hook event to its codex equivalent.
@@ -984,774 +878,5 @@ fn strip_hook_prose_from_codex_agents(global: bool, name: &str) -> Result<()> {
     Ok(())
 }
 
-/// Remove hook instructions and permission entries from OpenCode opencode.json
-fn remove_hook_from_opencode_json(global: bool, name: &str) -> Result<()> {
-    validate_item_name(name)?;
-    let config_path = if global {
-        crate::config::opencode_global_config_path()
-    } else {
-        crate::config::opencode_project_config_path()
-    };
-    let instruction_path = opencode_hook_instruction_path(global, name);
-    let instruction_ref = opencode_hook_instruction_ref(global, name);
-    remove_hook_from_opencode_json_at_path(&config_path, &instruction_path, &instruction_ref, name)
-}
-
-fn remove_hook_from_opencode_json_at_path(
-    config_path: &Path,
-    instruction_path: &Path,
-    instruction_ref: &str,
-    name: &str,
-) -> Result<()> {
-    validate_item_name(name)?;
-    if let Some(parent) = instruction_path.parent()
-        && parent.exists()
-    {
-        let file_name = instruction_path
-            .file_name()
-            .and_then(|name| name.to_str())
-            .context("OpenCode hook instruction path missing file name")?;
-        checked_child_path(parent, file_name)?;
-    }
-    if !config_path.exists() {
-        let _ = std::fs::remove_file(instruction_path);
-        return Ok(());
-    }
-    let content = std::fs::read_to_string(config_path)?;
-    let mut config: serde_json::Value = serde_json::from_str(&content)?;
-
-    let mut changed = false;
-
-    // Remove the current file-path based format plus the legacy inline prose format.
-    let keywords: Vec<&str> = name.split('-').collect();
-    if let Some(instructions) = config
-        .get_mut("instructions")
-        .and_then(|i| i.as_array_mut())
-    {
-        let before = instructions.len();
-        instructions.retain(|i| {
-            let Some(s) = i.as_str() else { return true };
-            if s == instruction_ref {
-                return false;
-            }
-            let s_lower = s.to_lowercase();
-            !keywords.iter().all(|kw| s_lower.contains(kw))
-        });
-        if instructions.len() != before {
-            changed = true;
-        }
-    }
-
-    let remove_instruction = instruction_path.exists();
-
-    // If no vstack hook instructions remain, remove the temporary bash restriction we added.
-    if let Some(map) = config.as_object_mut() {
-        let no_vstack_hook_instructions = map
-            .get("instructions")
-            .and_then(|i| i.as_array())
-            .is_none_or(|entries| {
-                !entries.iter().any(|entry| {
-                    entry
-                        .as_str()
-                        .is_some_and(|value| value.contains("vstack-hook-"))
-                })
-            });
-
-        if let Some(instructions) = map.get("instructions").and_then(|i| i.as_array())
-            && instructions.is_empty()
-        {
-            map.remove("instructions");
-            changed = true;
-        }
-
-        if no_vstack_hook_instructions
-            && let Some(permission) = map.get_mut("permission").and_then(|p| p.as_object_mut())
-        {
-            let remove_bash = permission
-                .get("bash")
-                .and_then(|bash| bash.as_object())
-                .is_some_and(|bash| {
-                    bash.len() == 1
-                        && bash
-                            .get("*")
-                            .and_then(|value| value.as_str())
-                            .is_some_and(|value| value == "ask")
-                });
-            if remove_bash {
-                permission.remove("bash");
-                changed = true;
-            }
-            if permission.is_empty() {
-                map.remove("permission");
-                changed = true;
-            }
-        }
-    }
-
-    if changed {
-        let output = serde_json::to_string_pretty(&config)?;
-        std::fs::write(config_path, output)?;
-    }
-    if remove_instruction {
-        let _ = std::fs::remove_file(instruction_path);
-    }
-    Ok(())
-}
-
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn hook_fixture(name: &str, event: &str, matcher: Option<&str>) -> Hook {
-        Hook {
-            name: name.into(),
-            event: event.into(),
-            matcher: matcher.map(|m| m.into()),
-            description: format!("{name} test hook"),
-            safety: None,
-            timeout: Some(30),
-            harnesses: None,
-            script: format!("#!/usr/bin/env bash\n# {name}\nexit 0\n"),
-            source_path: PathBuf::new(),
-        }
-    }
-
-    fn tmpdir(label: &str) -> PathBuf {
-        let dir = std::env::temp_dir().join(format!(
-            "vstack_{label}_{}_{}",
-            std::process::id(),
-            crate::config::now_iso().replace([':', '-'], "")
-        ));
-        let _ = std::fs::remove_dir_all(&dir);
-        std::fs::create_dir_all(&dir).unwrap();
-        dir
-    }
-
-    #[test]
-    fn codex_event_for_known_events() {
-        assert_eq!(codex_event_for("PreToolUse"), Some("PreToolUse"));
-        assert_eq!(codex_event_for("PostToolUse"), Some("PostToolUse"));
-        assert_eq!(codex_event_for("Stop"), Some("Stop"));
-        assert_eq!(codex_event_for("SessionStart"), Some("SessionStart"));
-    }
-
-    #[test]
-    fn codex_event_for_taskcompleted_is_unmapped() {
-        // TaskCompleted has no clean codex equivalent — routes to prose fallback.
-        assert_eq!(codex_event_for("TaskCompleted"), None);
-    }
-
-    #[test]
-    fn merge_codex_hooks_json_creates_new_file() {
-        let dir = tmpdir("codex_merge_new");
-        let hooks_json = dir.join("hooks.json");
-        let hook = hook_fixture("block-bare-cd", "PreToolUse", Some("Bash"));
-        let command = "bash /tmp/block-bare-cd.sh";
-        merge_codex_hooks_json(&hooks_json, "PreToolUse", &hook, command).unwrap();
-
-        let doc: serde_json::Value =
-            serde_json::from_str(&std::fs::read_to_string(&hooks_json).unwrap()).unwrap();
-        let arr = doc
-            .pointer("/hooks/PreToolUse")
-            .and_then(|v| v.as_array())
-            .expect("PreToolUse array present");
-        assert_eq!(arr.len(), 1);
-        assert_eq!(
-            arr[0].pointer("/matcher").and_then(|v| v.as_str()),
-            Some("Bash")
-        );
-        assert_eq!(
-            arr[0].pointer("/hooks/0/command").and_then(|v| v.as_str()),
-            Some(command)
-        );
-        assert_eq!(
-            arr[0].pointer("/hooks/0/timeout").and_then(|v| v.as_u64()),
-            Some(30)
-        );
-        let _ = std::fs::remove_dir_all(&dir);
-    }
-
-    #[test]
-    fn merge_codex_hooks_json_is_idempotent() {
-        let dir = tmpdir("codex_merge_idempotent");
-        let hooks_json = dir.join("hooks.json");
-        let hook = hook_fixture("block-bare-cd", "PreToolUse", Some("Bash"));
-        let command = "bash /tmp/block-bare-cd.sh";
-        merge_codex_hooks_json(&hooks_json, "PreToolUse", &hook, command).unwrap();
-        merge_codex_hooks_json(&hooks_json, "PreToolUse", &hook, command).unwrap();
-        let doc: serde_json::Value =
-            serde_json::from_str(&std::fs::read_to_string(&hooks_json).unwrap()).unwrap();
-        assert_eq!(
-            doc.pointer("/hooks/PreToolUse")
-                .and_then(|v| v.as_array())
-                .map(|a| a.len()),
-            Some(1)
-        );
-        let _ = std::fs::remove_dir_all(&dir);
-    }
-
-    #[test]
-    fn merge_codex_hooks_json_replaces_existing_hook_registration() {
-        let dir = tmpdir("codex_merge_replace");
-        let hooks_json = dir.join("hooks.json");
-        std::fs::write(
-            &hooks_json,
-            r#"{
-  "hooks": {
-    "PreToolUse": [
-      {
-        "matcher": "Bash",
-        "hooks": [{"type": "command", "command": "bash /home/.codex/hooks/guard.sh", "timeout": 30}]
-      }
-    ]
-  }
-}"#,
-        )
-        .unwrap();
-        let mut hook = hook_fixture("guard", "PostCompact", None);
-        hook.timeout = Some(5);
-        merge_codex_hooks_json(
-            &hooks_json,
-            "PostCompact",
-            &hook,
-            "bash /home/.codex/hooks/guard.sh",
-        )
-        .unwrap();
-
-        let doc: serde_json::Value =
-            serde_json::from_str(&std::fs::read_to_string(&hooks_json).unwrap()).unwrap();
-        assert!(doc.pointer("/hooks/PreToolUse").is_none());
-        let arr = doc
-            .pointer("/hooks/PostCompact")
-            .and_then(|v| v.as_array())
-            .expect("PostCompact array present");
-        assert_eq!(arr.len(), 1);
-        assert!(arr[0].pointer("/matcher").is_none());
-        assert_eq!(
-            arr[0].pointer("/hooks/0/timeout").and_then(|v| v.as_u64()),
-            Some(5)
-        );
-        let _ = std::fs::remove_dir_all(&dir);
-    }
-
-    #[test]
-    fn validate_item_name_rejects_path_like_names() {
-        for name in [
-            "",
-            ".",
-            "..",
-            "../victim",
-            "a/b",
-            "a\\b",
-            "/abs",
-            "-leading-dash",
-            "bad\";touch pwn;#",
-            "bad$(touch pwn)",
-            "has spaces",
-            "has\nnewline",
-            "bad`touch pwn`",
-        ] {
-            assert!(validate_item_name(name).is_err(), "accepted {name:?}");
-        }
-        assert!(validate_item_name("guard-hook").is_ok());
-        assert!(validate_item_name("guard.hook_1").is_ok());
-    }
-
-    #[test]
-    fn hook_prune_preserves_user_handlers_with_same_basename() {
-        let mut hooks_obj = serde_json::json!({
-            "PreToolUse": [
-                {
-                    "matcher": "Bash",
-                    "hooks": [{"type": "command", "command": "bash ./scripts/guard.sh"}]
-                },
-                {
-                    "matcher": "Bash",
-                    "hooks": [{"type": "command", "command": "bash /usr/local/bin/guard.sh"}]
-                },
-                {
-                    "matcher": "Bash",
-                    "hooks": [{"type": "command", "command": "bash \"$CLAUDE_PROJECT_DIR/.claude/hooks/guard.sh\""}]
-                }
-            ]
-        })
-        .as_object()
-        .unwrap()
-        .clone();
-        let owned = vec!["bash \"$CLAUDE_PROJECT_DIR/.claude/hooks/guard.sh\"".to_string()];
-
-        assert!(remove_hook_entries_from_hooks_object(
-            &mut hooks_obj,
-            &owned
-        ));
-        let arr = hooks_obj
-            .get("PreToolUse")
-            .and_then(|v| v.as_array())
-            .unwrap();
-        assert_eq!(arr.len(), 2, "only vstack-owned command should be pruned");
-        let body = serde_json::to_string(&hooks_obj).unwrap();
-        assert!(body.contains("./scripts/guard.sh"));
-        assert!(body.contains("/usr/local/bin/guard.sh"));
-        assert!(!body.contains(".claude/hooks/guard.sh"));
-    }
-
-    #[test]
-    fn merge_codex_hooks_json_does_not_dedupe_substring_collisions() {
-        // A hook named `foo` must not be considered already-present when the
-        // event already has `notfoo.sh`; only exact vstack-owned commands are
-        // pruned.
-        let dir = tmpdir("codex_merge_substring");
-        let hooks_json = dir.join("hooks.json");
-        std::fs::write(
-            &hooks_json,
-            r#"{
-  "hooks": {
-    "PreToolUse": [
-      {
-        "matcher": "Bash",
-        "hooks": [{"type": "command", "command": "bash /home/.codex/hooks/notfoo.sh"}]
-      }
-    ]
-  }
-}"#,
-        )
-        .unwrap();
-        let hook = hook_fixture("foo", "PreToolUse", Some("Bash"));
-        merge_codex_hooks_json(
-            &hooks_json,
-            "PreToolUse",
-            &hook,
-            "bash /home/.codex/hooks/foo.sh",
-        )
-        .unwrap();
-
-        let doc: serde_json::Value =
-            serde_json::from_str(&std::fs::read_to_string(&hooks_json).unwrap()).unwrap();
-        let arr = doc
-            .pointer("/hooks/PreToolUse")
-            .and_then(|v| v.as_array())
-            .unwrap();
-        assert_eq!(
-            arr.len(),
-            2,
-            "`foo.sh` must not collide with existing `notfoo.sh`"
-        );
-    }
-
-    #[test]
-    fn merge_codex_hooks_json_preserves_existing_entries() {
-        let dir = tmpdir("codex_merge_preserve");
-        let hooks_json = dir.join("hooks.json");
-        std::fs::write(
-            &hooks_json,
-            r#"{
-  "hooks": {
-    "PreToolUse": [
-      {
-        "matcher": "Bash",
-        "hooks": [{"type": "command", "command": "bash /user/own.sh"}]
-      }
-    ]
-  }
-}"#,
-        )
-        .unwrap();
-
-        let hook = hook_fixture("new-one", "PreToolUse", Some("Bash"));
-        merge_codex_hooks_json(&hooks_json, "PreToolUse", &hook, "bash /tmp/new-one.sh").unwrap();
-
-        let doc: serde_json::Value =
-            serde_json::from_str(&std::fs::read_to_string(&hooks_json).unwrap()).unwrap();
-        let arr = doc
-            .pointer("/hooks/PreToolUse")
-            .and_then(|v| v.as_array())
-            .unwrap();
-        assert_eq!(arr.len(), 2, "user entry should be preserved");
-        let _ = std::fs::remove_dir_all(&dir);
-    }
-
-    #[test]
-    fn merge_codex_hooks_json_preserves_user_handler_with_same_basename() {
-        let dir = tmpdir("codex_merge_preserve_same_basename");
-        let hooks_json = dir.join("hooks.json");
-        std::fs::write(
-            &hooks_json,
-            r#"{
-  "hooks": {
-    "PreToolUse": [
-      {
-        "matcher": "Bash",
-        "hooks": [{"type": "command", "command": "bash /usr/local/bin/guard.sh"}]
-      }
-    ]
-  }
-}"#,
-        )
-        .unwrap();
-
-        let hook = hook_fixture("guard", "PreToolUse", Some("Bash"));
-        merge_codex_hooks_json(&hooks_json, "PreToolUse", &hook, "bash /tmp/guard.sh").unwrap();
-
-        let doc: serde_json::Value =
-            serde_json::from_str(&std::fs::read_to_string(&hooks_json).unwrap()).unwrap();
-        let arr = doc
-            .pointer("/hooks/PreToolUse")
-            .and_then(|v| v.as_array())
-            .unwrap();
-        assert_eq!(arr.len(), 2, "user entry with same basename should remain");
-        let body = serde_json::to_string(&doc).unwrap();
-        assert!(body.contains("/usr/local/bin/guard.sh"));
-        assert!(body.contains("/tmp/guard.sh"));
-        let _ = std::fs::remove_dir_all(&dir);
-    }
-
-    #[test]
-    fn remove_hook_install_codex_strips_script_json_and_legacy_prose() {
-        let dir = tmpdir("codex_remove_strip");
-        let hooks_dir = dir.join("hooks");
-        let agents_dir = dir.join("agents");
-        std::fs::create_dir_all(&hooks_dir).unwrap();
-        std::fs::create_dir_all(&agents_dir).unwrap();
-        std::fs::write(hooks_dir.join("post-edit-lint.sh"), "#!/usr/bin/env bash\n").unwrap();
-        std::fs::write(hooks_dir.join("block-bare-cd.sh"), "#!/usr/bin/env bash\n").unwrap();
-        let post_edit_command = format!("bash {}", hooks_dir.join("post-edit-lint.sh").display());
-        let block_bare_command = format!("bash {}", hooks_dir.join("block-bare-cd.sh").display());
-
-        let hooks_json = dir.join("hooks.json");
-        let hooks_doc = serde_json::json!({
-            "hooks": {
-                "PreToolUse": [
-                    {
-                        "matcher": "Bash",
-                        "hooks": [{"type": "command", "command": block_bare_command}]
-                    },
-                    {
-                        "matcher": "Bash",
-                        "hooks": [{"type": "command", "command": "bash /home/.codex/hooks/user-own.sh"}]
-                    }
-                ],
-                "PostToolUse": [
-                    {
-                        "matcher": "Edit|Write",
-                        "hooks": [{"type": "command", "command": post_edit_command}]
-                    }
-                ]
-            }
-        });
-        std::fs::write(
-            &hooks_json,
-            serde_json::to_string_pretty(&hooks_doc).unwrap(),
-        )
-        .unwrap();
-        let agent_toml = agents_dir.join("rust.toml");
-        std::fs::write(
-            &agent_toml,
-            r#"name = "rust"
-developer_instructions = '''
-Body
-
-## Safety: post-edit-lint
-
-Remove me.
-
-## Keep
-
-Keep me.
-'''
-"#,
-        )
-        .unwrap();
-
-        crate::test_util::with_codex_home(&dir, || {
-            remove_hook_install("post-edit-lint", Harness::Codex, true).unwrap();
-        });
-
-        let result: serde_json::Value =
-            serde_json::from_str(&std::fs::read_to_string(&hooks_json).unwrap()).unwrap();
-        assert!(!hooks_dir.join("post-edit-lint.sh").exists());
-        assert!(hooks_dir.join("block-bare-cd.sh").exists());
-        assert!(
-            result.pointer("/hooks/PostToolUse").is_none(),
-            "empty PostToolUse should be pruned"
-        );
-        let pre = result
-            .pointer("/hooks/PreToolUse")
-            .unwrap()
-            .as_array()
-            .unwrap();
-        assert_eq!(pre.len(), 2, "unrelated PreToolUse entries preserved");
-        let agent = std::fs::read_to_string(agent_toml).unwrap();
-        assert!(!agent.contains("Safety: post-edit-lint"));
-        assert!(agent.contains("## Keep"));
-        let _ = std::fs::remove_dir_all(&dir);
-    }
-
-    #[test]
-    fn enable_codex_hooks_feature_creates_section() {
-        let dir = tmpdir("codex_features_new");
-        let config = dir.join("config.toml");
-        enable_codex_hooks_feature(&config).unwrap();
-        let body = std::fs::read_to_string(&config).unwrap();
-        assert!(body.contains("[features]"));
-        assert!(body.contains("hooks = true"));
-        assert!(!body.contains("codex_hooks"));
-        let _ = std::fs::remove_dir_all(&dir);
-    }
-
-    #[test]
-    fn enable_codex_hooks_feature_is_idempotent() {
-        let dir = tmpdir("codex_features_idempotent");
-        let config = dir.join("config.toml");
-        enable_codex_hooks_feature(&config).unwrap();
-        let body1 = std::fs::read_to_string(&config).unwrap();
-        enable_codex_hooks_feature(&config).unwrap();
-        let body2 = std::fs::read_to_string(&config).unwrap();
-        assert_eq!(body1, body2);
-        let _ = std::fs::remove_dir_all(&dir);
-    }
-
-    #[test]
-    fn enable_codex_hooks_feature_preserves_user_content() {
-        let dir = tmpdir("codex_features_preserve");
-        let config = dir.join("config.toml");
-        std::fs::write(
-            &config,
-            "# user comment\nmodel = \"gpt-5.5\"\n\n[other]\nfoo = 1\n",
-        )
-        .unwrap();
-        enable_codex_hooks_feature(&config).unwrap();
-        let body = std::fs::read_to_string(&config).unwrap();
-        assert!(body.contains("# user comment"));
-        assert!(body.contains("model = \"gpt-5.5\""));
-        assert!(body.contains("[other]"));
-        assert!(body.contains("hooks = true"));
-        assert!(!body.contains("codex_hooks"));
-        let _ = std::fs::remove_dir_all(&dir);
-    }
-
-    #[test]
-    fn enable_codex_hooks_feature_inserts_under_existing_features() {
-        let dir = tmpdir("codex_features_existing");
-        let config = dir.join("config.toml");
-        std::fs::write(
-            &config,
-            "[features]\nother_flag = true\n\n[unrelated]\nx = 1\n",
-        )
-        .unwrap();
-        enable_codex_hooks_feature(&config).unwrap();
-        let body = std::fs::read_to_string(&config).unwrap();
-        let features_pos = body.find("[features]").unwrap();
-        let unrelated_pos = body.find("[unrelated]").unwrap();
-        let hooks_pos = body.find("hooks = true").unwrap();
-        assert!(features_pos < hooks_pos && hooks_pos < unrelated_pos);
-        let _ = std::fs::remove_dir_all(&dir);
-    }
-
-    #[test]
-    fn enable_codex_hooks_feature_migrates_deprecated_flag() {
-        let dir = tmpdir("codex_features_migrate");
-        let config = dir.join("config.toml");
-        std::fs::write(
-            &config,
-            "[features]\ncodex_hooks = true\nother_flag = true\n\n[unrelated]\nx = 1\n",
-        )
-        .unwrap();
-        enable_codex_hooks_feature(&config).unwrap();
-        let body = std::fs::read_to_string(&config).unwrap();
-        assert!(body.contains("hooks = true"));
-        assert!(body.contains("other_flag = true"));
-        assert!(!body.contains("codex_hooks"));
-        let hooks_pos = body.find("hooks = true").unwrap();
-        let unrelated_pos = body.find("[unrelated]").unwrap();
-        assert!(hooks_pos < unrelated_pos);
-        let _ = std::fs::remove_dir_all(&dir);
-    }
-
-    #[test]
-    fn migrate_codex_hooks_feature_does_not_create_config() {
-        let dir = tmpdir("codex_features_migrate_no_config");
-        let config = dir.join("config.toml");
-        migrate_codex_hooks_feature(&config).unwrap();
-        assert!(!config.exists());
-        let _ = std::fs::remove_dir_all(&dir);
-    }
-
-    #[test]
-    fn migrate_codex_hooks_feature_preserves_deprecated_value() {
-        let dir = tmpdir("codex_features_migrate_only");
-        let config = dir.join("config.toml");
-        std::fs::write(
-            &config,
-            "[features]\ncodex_hooks = false\nother_flag = true\n\n[unrelated]\nx = 1\n",
-        )
-        .unwrap();
-        migrate_codex_hooks_feature(&config).unwrap();
-        let body = std::fs::read_to_string(&config).unwrap();
-        assert!(body.contains("hooks = false"));
-        assert!(body.contains("other_flag = true"));
-        assert!(!body.contains("codex_hooks"));
-        let hooks_pos = body.find("hooks = false").unwrap();
-        let unrelated_pos = body.find("[unrelated]").unwrap();
-        assert!(hooks_pos < unrelated_pos);
-        let _ = std::fs::remove_dir_all(&dir);
-    }
-
-    #[test]
-    fn migrate_codex_hooks_feature_prefers_existing_hooks_value() {
-        let dir = tmpdir("codex_features_migrate_existing_hooks");
-        let config = dir.join("config.toml");
-        std::fs::write(
-            &config,
-            "[features]\nhooks = false\ncodex_hooks = true\nother_flag = true\n",
-        )
-        .unwrap();
-        migrate_codex_hooks_feature(&config).unwrap();
-        let body = std::fs::read_to_string(&config).unwrap();
-        assert!(body.contains("hooks = false"));
-        assert!(!body.contains("hooks = true"));
-        assert!(!body.contains("codex_hooks"));
-        let _ = std::fs::remove_dir_all(&dir);
-    }
-
-    #[test]
-    fn enable_codex_hooks_feature_replaces_disabled_hooks_flag() {
-        let dir = tmpdir("codex_features_disabled");
-        let config = dir.join("config.toml");
-        std::fs::write(&config, "[features]\nhooks = false\ncodex_hooks = true\n").unwrap();
-        enable_codex_hooks_feature(&config).unwrap();
-        let body = std::fs::read_to_string(&config).unwrap();
-        assert!(body.contains("hooks = true"));
-        assert!(!body.contains("hooks = false"));
-        assert!(!body.contains("codex_hooks"));
-        let _ = std::fs::remove_dir_all(&dir);
-    }
-
-    #[test]
-    fn remove_hook_from_opencode_removes_instruction() {
-        let base = std::env::temp_dir().join("vstack_test_opencode");
-        let _ = std::fs::create_dir_all(&base);
-        let config_path = base.join("opencode.json");
-        let instruction_path = base
-            .join(".opencode")
-            .join("instructions")
-            .join("vstack-hook-block-bare-cd.md");
-        std::fs::create_dir_all(instruction_path.parent().unwrap()).unwrap();
-        std::fs::write(&instruction_path, "# Safety").unwrap();
-
-        let content = r#"{
-  "$schema": "https://opencode.ai/config.json",
-  "instructions": [
-    ".opencode/instructions/vstack-hook-block-bare-cd.md"
-  ],
-  "permission": {
-    "bash": {
-      "*": "ask"
-    }
-  }
-}"#;
-        std::fs::write(&config_path, content).unwrap();
-
-        remove_hook_from_opencode_json_at_path(
-            &config_path,
-            &instruction_path,
-            ".opencode/instructions/vstack-hook-block-bare-cd.md",
-            "block-bare-cd",
-        )
-        .unwrap();
-
-        let result = std::fs::read_to_string(&config_path).unwrap();
-        let parsed: serde_json::Value = serde_json::from_str(&result).unwrap();
-
-        // instructions and permission should be gone
-        assert!(
-            parsed.get("instructions").is_none(),
-            "instructions should be removed, got: {result}"
-        );
-        assert!(
-            parsed.get("permission").is_none(),
-            "permission should be removed, got: {result}"
-        );
-        assert!(
-            !instruction_path.exists(),
-            "instruction file should be removed"
-        );
-
-        // Cleanup
-        let _ = std::fs::remove_dir_all(&base);
-    }
-
-    #[test]
-    fn remove_hook_from_opencode_keeps_instruction_when_config_parse_fails() {
-        let base = std::env::temp_dir().join("vstack_test_opencode_invalid_config");
-        let _ = std::fs::remove_dir_all(&base);
-        std::fs::create_dir_all(&base).unwrap();
-        let config_path = base.join("opencode.json");
-        let instruction_path = base.join("instructions").join("vstack-hook-guard.md");
-        std::fs::create_dir_all(instruction_path.parent().unwrap()).unwrap();
-        std::fs::write(&instruction_path, "# Safety").unwrap();
-        std::fs::write(&config_path, "{not-json").unwrap();
-
-        let result = remove_hook_from_opencode_json_at_path(
-            &config_path,
-            &instruction_path,
-            "instructions/vstack-hook-guard.md",
-            "guard",
-        );
-
-        assert!(result.is_err());
-        assert!(
-            instruction_path.exists(),
-            "instruction file should remain when config cleanup fails"
-        );
-        let _ = std::fs::remove_dir_all(&base);
-    }
-
-    #[test]
-    fn remove_hook_from_opencode_preserves_unrelated_permissions() {
-        let base = std::env::temp_dir().join("vstack_test_opencode_permissions");
-        let _ = std::fs::create_dir_all(&base);
-        let config_path = base.join("opencode.json");
-        let instruction_path = base.join("instructions").join("vstack-hook-review-bash.md");
-        std::fs::create_dir_all(instruction_path.parent().unwrap()).unwrap();
-        std::fs::write(&instruction_path, "# Safety").unwrap();
-
-        let content = r#"{
-  "$schema": "https://opencode.ai/config.json",
-  "instructions": [
-    "instructions/vstack-hook-review-bash.md"
-  ],
-  "permission": {
-    "edit": "deny",
-    "bash": {
-      "*": "ask"
-    }
-  }
-}"#;
-        std::fs::write(&config_path, content).unwrap();
-
-        remove_hook_from_opencode_json_at_path(
-            &config_path,
-            &instruction_path,
-            "instructions/vstack-hook-review-bash.md",
-            "review-bash",
-        )
-        .unwrap();
-
-        let result = std::fs::read_to_string(&config_path).unwrap();
-        let parsed: serde_json::Value = serde_json::from_str(&result).unwrap();
-
-        assert_eq!(
-            parsed.get("permission").and_then(|p| p.get("edit")),
-            Some(&serde_json::Value::String("deny".into()))
-        );
-        assert!(
-            parsed
-                .get("permission")
-                .and_then(|p| p.get("bash"))
-                .is_none(),
-            "vstack-added bash permission should be removed, got: {result}"
-        );
-
-        let _ = std::fs::remove_dir_all(&base);
-    }
-}
+mod tests;
